refactor: 각각의 역할에 맞게 패키지 분리
domain.rule - 발판 생성 규칙
