+++ src/main/java/laddergame/LadderGameApplication.java
@@ -2,6 +2,8 @@ package laddergame;
 
 import laddergame.controller.Maker;
 import laddergame.domain.*;
+import laddergame.domain.rule.RandomCreateRule;
+import laddergame.domain.rule.Rule;
 import laddergame.util.InputView;
 import laddergame.util.OutputView;
 
@@ -10,7 +12,7 @@ public class LadderGameApplication {
         Tags members = Maker.makeMembers();
         Tags prizes = Maker.makePrizes(members.size());
         int height = InputView.inputHeight();
-        Rule rule = new RandomCreate();
+        Rule rule = new RandomCreateRule();
         Ladder ladder = LadderGenerator.generateLadder(height, members.size(), rule);
         LadderGameResult ladderGameResult = LadderGame.startGame(members, ladder, prizes);
