--- src/main/java/laddergame/domain/AlwaysCreate.java
@@ -1,8 +1,0 @@
-package laddergame.domain;
-
-public class AlwaysCreate implements Rule {
-    @Override
-    public boolean canCreate() {
-        return true;
-    }
-}
