+++ src/main/java/laddergame/domain/LadderGenerator.java
@@ -1,5 +1,7 @@
 package laddergame.domain;
 
+import laddergame.domain.rule.Rule;
+
 public class LadderGenerator {
     public static Ladder generateLadder(int height, int numberOfPerson, Rule rule) {
         Ladder ladder = new Ladder();
