+++ src/main/java/laddergame/domain/rule/AlwaysCreateRule.java
@@ -1,0 +1,8 @@
+package laddergame.domain.rule;
+
+public class AlwaysCreateRule implements Rule {
+    @Override
+    public boolean canCreate() {
+        return true;
+    }
+}
