+++ src/main/java/laddergame/domain/rule/RandomCreateRule.java
@@ -1,8 +1,8 @@
-package laddergame.domain;
+package laddergame.domain.rule;
 
 import laddergame.util.RandomGenerator;
 
-public class RandomCreate implements Rule {
+public class RandomCreateRule implements Rule {
     @Override
     public boolean canCreate() {
         return RandomGenerator.makeTrueOrFalse();
