+++ src/main/java/laddergame/domain/rule/Rule.java
@@ -1,4 +1,4 @@
-package laddergame.domain;
+package laddergame.domain.rule;
 
 public interface Rule {
     boolean canCreate();
