+++ src/test/java/laddergame/domain/LadderGameTest.java
@@ -1,5 +1,6 @@
 package laddergame.domain;
 
+import laddergame.domain.rule.AlwaysCreateRule;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
@@ -9,7 +10,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 class LadderGameTest {
     @Test
     void 게임_결과가_올바른지_테스트() {
-        Ladder ladder = LadderGenerator.generateLadder(5, 2, new AlwaysCreate());
+        Ladder ladder = LadderGenerator.generateLadder(5, 2, new AlwaysCreateRule());
         Tags members = new Tags(Arrays.asList("pobi", "cu"));
         Tags prizes = new Tags(Arrays.asList("1000", "100"));
         LadderGameResult result = LadderGame.startGame(members, ladder, prizes);
