+++ src/test/java/laddergame/domain/LadderTest.java
@@ -1,5 +1,6 @@
 package laddergame.domain;
 
+import laddergame.domain.rule.AlwaysCreateRule;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
@@ -17,7 +18,7 @@ class LadderTest {
 
     @Test
     void 사다리를_타고_제대로_내려가는지_확인() {
-        Ladder ladder = LadderGenerator.generateLadder(5, 2, new AlwaysCreate());
+        Ladder ladder = LadderGenerator.generateLadder(5, 2, new AlwaysCreateRule());
 
         assertThat(ladder.takeLadder(0)).isEqualTo(1);
         assertThat(ladder.takeLadder(1)).isEqualTo(0);
