+++ src/test/java/laddergame/domain/LineGeneratorTest.java
@@ -1,5 +1,6 @@
 package laddergame.domain;
 
+import laddergame.domain.rule.AlwaysCreateRule;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
@@ -9,7 +10,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 class LineGeneratorTest {
     @Test
     void Line이_제대로_생성되는지_테스트() {
-        assertThat(LineGenerator.lineGenerate(2, new AlwaysCreate())).isEqualTo(
+        assertThat(LineGenerator.lineGenerate(2, new AlwaysCreateRule())).isEqualTo(
                 new Line(Arrays.asList(false, true, false)));
     }
 }
